+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+# 116.0.0
+
+In this major release, we deleted the `rippleAllowsSelection` property from MDCChipView and added containerRadius support to filled text fields.
+
+## API changes
+
+`rippleAllowsSelection` has been deleted from MDCChipView.
+
+## Component changes
+
+### Chips
+
+* [Have Ripple imitate Ink's behavior in MDCChipView.](https://github.com/material-components/material-components-ios/commit/6543af21be74965abb609e792d744f8f76fe4c63) (Yarden Eitan)
+
+## Multi-component changes
+
+* [Add containerRadius support to filled text field](https://github.com/material-components/material-components-ios/commit/0907c196231629d9af7e86feeed86748ec96a20f) (Andrew Overton)
+* [Internal change](https://github.com/material-components/material-components-ios/commit/9f62e64603adc68475ee3d21981dd79cdaff1129) (Yarden Eitan)
+
+---
+
 # 115.1.0
 
 In this minor release we add 2 APIs to customize BottomNavigation's appearance and behavior, we deprecate the property `rippleAllowsSelection` from `MDCChipView`, we provide a `containerRadius` API to `MDCBaseTextField` and `MDCBaseTextArea`, along with other code cleanups.